+++ tests/testthat/test-Dominance_Stats.r
@@ -35,7 +35,7 @@ cyl_c <- c(cyl_mgn[[1]], mean(cyl_mgn[2:3]), cyl_mgn[[4]])
 carb_c <- c(carb_mgn[[1]], mean(carb_mgn[2:3]), carb_mgn[[4]])
 
 cdl_names <- list(c("vs", "cyl", "carb"), paste0("IVs_", 1:3))
-cdl_names_new <- list(c("vs", "cyl", "carb"), paste0("subset_size_", 1:3))
+cdl_names_new <- list(c("vs", "cyl", "carb"), paste0("include_at_", 1:3))
 
 cdl_test <- matrix(c(vs_c, cyl_c, carb_c), nrow = 3, ncol = 3, 
                    byrow = TRUE, dimnames = cdl_names)
@@ -163,7 +163,7 @@ vs_fl_c <- c(vs_fl_mgn[[1]], mean(vs_fl_mgn[2:3]), vs_fl_mgn[[4]])
 cyl_fl_c <- c(cyl_fl_mgn[[1]], mean(cyl_fl_mgn[2:3]), cyl_fl_mgn[[4]])
 carb_fl_c <- c(carb_fl_mgn[[1]], mean(carb_fl_mgn[2:3]), carb_fl_mgn[[4]])
 
-cdl_fl_names_new <- list(c("mpg~vs", "mpg~cyl", "qsec~carb"), paste0("subset_size_", 1:3))
+cdl_fl_names_new <- list(c("mpg~vs", "mpg~cyl", "qsec~carb"), paste0("include_at_", 1:3))
 
 cdl_fl_test_new <- matrix(c(vs_fl_c, cyl_fl_c, carb_fl_c), nrow = 3, ncol = 3, 
                        byrow = TRUE, dimnames = cdl_fl_names_new)